requirements.txt
asgiref==3.8.1
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
djongo5==1.3.9
PyJWT==2.10.1
pymongo==3.13.0
pytz==2022.6
sqlparse==0.2.4
